perf(playground-ui): defer tool diff highlighting
