+++ packages/playground-ui/src/ds/components/Code/code.tsx
@@ -1,31 +1,11 @@
 import * as React from 'react';
-import type { ThemedToken } from 'shiki/core';
 
-import { highlight } from '../CodeEditor/highlight';
+import { HighlightedTokenLine } from './highlighted-code';
+import { useHighlightedCode } from './use-highlighted-code';
 
 export interface CodeProps extends React.HTMLAttributes<HTMLPreElement> {
   code: string;
   lang?: string;
-}
-
-function tokenStyle(token: ThemedToken): React.CSSProperties | undefined {
-  if (token.htmlStyle && typeof token.htmlStyle === 'object') {
-    return token.htmlStyle as React.CSSProperties;
-  }
-
-  return token.color ? { color: token.color } : undefined;
-}
-
-interface Highlighted {
-  code: string;
-  lang: string;
-  tokens: ThemedToken[][];
-}
-
-/** Colors from an earlier pass still hold when the new code only appends to the old. */
-function usableHighlight(highlighted: Highlighted | null, code: string, lang?: string): Highlighted | null {
-  if (!highlighted || highlighted.lang !== lang) return null;
-  return code.startsWith(highlighted.code) ? highlighted : null;
 }
 
 /**
@@ -42,28 +22,7 @@ function usableHighlight(highlighted: Highlighted | null, code: string, lang?: s
  * the newly arrived tail waits, uncolored, for the next pass.
  */
 export const Code = React.memo(function Code({ code, lang, ...props }: CodeProps) {
-  const [highlighted, setHighlighted] = React.useState<Highlighted | null>(null);
-
-  React.useEffect(() => {
-    if (!lang) {
-      setHighlighted(null);
-      return;
-    }
-
-    let cancelled = false;
-
-    void highlight(code, lang)
-      .then(tokens => {
-        if (!cancelled && tokens?.length) setHighlighted({ code, lang, tokens });
-      })
-      .catch(() => {});
-
-    return () => {
-      cancelled = true;
-    };
-  }, [code, lang]);
-
-  const usable = usableHighlight(highlighted, code, lang);
+  const usable = useHighlightedCode(code, lang);
   if (!usable) {
     return <pre {...props}>{code}</pre>;
   }
@@ -76,23 +35,13 @@ export const Code = React.memo(function Code({ code, lang, ...props }: CodeProps
       <code>
         {usable.tokens.map((line, lineIndex) => {
           const lineOffset = codeOffset;
-          let tokenOffset = lineOffset;
-          const tokenSpans = line.map(token => {
-            const key = tokenOffset;
-            tokenOffset += token.content.length;
-
-            return (
-              <span key={key} className="shiki-token" style={tokenStyle(token)}>
-                {token.content}
-              </span>
-            );
-          });
-
-          codeOffset = tokenOffset + 1;
+          codeOffset += line.reduce((length, token) => length + token.content.length, 0) + 1;
 
           return (
             <React.Fragment key={lineOffset}>
-              <span>{tokenSpans}</span>
+              <span>
+                <HighlightedTokenLine tokens={line} />
+              </span>
               {lineIndex !== usable.tokens.length - 1 && '\n'}
             </React.Fragment>
           );
